requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
chromadb>=0.4.0
tqdm>=4.65.0
pydantic>=2.0.0
